pyproject.toml
[project]
name = "recipe-app-backend-25-04-08"
version = "0.1.0"
description = ""
authors = [
    {name = "elena7993",email = "elenalee7993@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
